Set up node
This sets up our baseline node structure, including linting, bundling,
and transpiling.

As is the way, we're using airbnb's linter and not using semicolons.

This does NOT set up things like sequelize or react, as those should be
included when we're working on tasks that would actually need them.

Issue #8