Create abstract crawl tooling
Crawlers take in a URL and spit out a list of URLS that need to be
scraped in more detail.  This commit sets up the object model for our
various crawlers and creates an example implementation.

This does not offer a good way to actually trigger crawls yet.

AbstractCrawler is intended to be an abstract class (though abstract
classes are not explicitly supported in javascript).  Specifically, it
requires an extended class to implement the crawlHandler method.

Issue #20