Add bull package for queues
[bull](https://www.npmjs.com/package/bull) is a package for queue
management. We'll be using it to keep track of the scrapes and crawls
that need to be processed.  The benefit of using a queue manager is that
it allows us to architect our system in a way that can be scaled up and
down over time.  We'll also benefit from the ability to store upcoming
tasks in redis (so if something restarts or breaks we won't lose the
scrape).

One thing about bull is that it integrates with a few UIs for monitoring
the queue.  This isn't an immediate need, but it may really come in
handy down the line when we want folks to be able to see what the
status of the alerts system is at any given moment.

We will be baking queue management into our abstract crawler (and
scraper).

Issue #20