Add models and migrations for claims and speakers
Here we’ve got Sequelize models and migrations for claims and speakers,
as well as a seed of each. After merging this, you’ll want to run
`yarn migrate` and `sequelize db:seed:all` (the latter only if you
actually want the seeds in your database, so, *not in production*).

By default, Sequelize camelCases tables and columns for consistency
with the models and model attributes themselves, but we like to follow
snake-case for SQL and camelCase for JavaScript. So there’s also some
stitching here to make that mapping work properly. Just be aware that
when referencing SQL field names (such as in migrations and seed
files), you’ll want to use the snake-cased names (e.g. `created_at`),
while using camelCase (`createdAt`) when referencing
Sequelize/JavaScript model names/attributes. It can be weird.

Affects #14

Co-authored-by: Asa Royal <asa@asaroyal.com>