Create crawl processors
As we move to a queue system for triggering crawls, "processor" files
capture the logic of actually kicking off a crawl and ultimately
handling the results.

We're adopting a new pattern of having the individual crawl sit in a
folder with a paired `processing.js`.

This leverages the [separate
processes](https://www.npmjs.com/package/bull#separate-processes)
feature of bull, which means we can orchestrate the recurring queues
themselves through a separate mechanism TBD.

Issue #20
Issue #43