Add CNN transcript portal crawler
CNN publishes transcripts on a continuous basis and has a few types of
transcript list.  Specifically they have a transcript list for each day,
show, and a special transcript list of "breaking news" transcripts.

All of these lists are linked to from a single transcript portal.

This commit creates the crawler of that portal.  It does not result in
a list of transcript URLs but rather it returns a list of transcript
list URLs which then need to be passed to a not-yet-written
CnnTranscriptListCrawler.

This also kicks off a utils/crawler.js tool which abstracts common
compound cheerio tasks that are not unique to specific crawlers.

It also makes two changes to the AbstractQueueScheduler:

1. It makes it possible for an extending class to pass in data to the
   job by overriding the getJobData method.
2. It removes the job name attribute, which is apparently more than just
   a friendly label and actually needs to be explicitly linked to the
   processor when the processor is registered.

Issue #24