Create configured logger
This makes a configured instance of a winston logger available for use
in the rest of the code base.  It does not replace instances of
console.log but the intent is that we can simply import this logger into
a file and use logger.* to write to the appropriate log destination.

All logs are written as json in a logs/* directory.  Errors are written
to two files (errors.log and consolidated.log) while other levels of
logging are only written to the consolidated log file.

In non-production environments the logs are also written to console.

Issue #34