Don’t wrap `Mailer#send()` in a redundant promise
Per observation by @slifty, `mailgun-js`’s `Mailer#send()` already
returns a promise, so my wrapping it in `Promise.resolve()` was
unnecessary.

In my defense it was like 1am and I’d been in Promises hell which is a
terrible defense and so I rest my case objection your honor.

Affects #22