Replace console logs/errors with logger
Since #55, we’ve had proper logging. 🎉 Replace uses of `console` with
the logger.

Also, I’m removing the TODO that I _just added_ for mailgun-js’s
`testModeLogger()`. It’s a weird beast: it’s a function that runs when
you’re in test mode (no actual interaction with the Mailgun API occurs)
so that you can choose how to log/inspect the request data.

Yet, we only want to use “test mode” when in literal test mode (like,
when running automated tests), at which point we don’t _want_ the
logger to log anything. We’d only use this during dev, but during dev
we _want_ the Mailgun API to be interacted with.

So anyway, I’m keeping this an essential noop that just collects the
useful request data and returns it but doesn’t log it. FIN. If we want
to noodle with this later, there it is.

Affects #22