Create common AbstractScraper
The AbstractCrawler and AbstractStatementScraper are both different
types of scraper, and they should both extend a common abstract class
so that we can more easily add common functionality to both.

This creates that common AbstractScraper, thoguh it is possible the
location of the AbstractScraper should be somewhere else (or we should
combine the crawlers / scrapers into a single directory)

This does not begin to add that functionality beyond creating a common
log info message to show off the potential.

Issue #60