Add scrape redis storage
We want to know when scrapes succeed and fail so that we can be more
intelligent about avoiding re-scrapes.  This commit bakes in the ability
to write and read this information through redis.  We decided to use
redis instead of SQL because ultimately this is just a glorified cache.
In the worst case the scrape log is lost and we re-scrape content that
was already scraped.

We store a few kinds of information for each scraped URL:

- (base) the timestamp of the most recent scrape attempt
- (successLedger) the list of timestamps of successful scrapes
- (failureLedger) the list of timestamps of failed scrapes
- (errors) the contents of errors (on failure)
- (results) the contents of results (on success)

Issue #66