Rename statement attribute
The statement scraper extracts a list of statement objects.  The object
is not explicitly codified (this should change someday) and as a result
we didn't notice that the statement object has a statement attribute.

Really, statements are a combination of a speaker and some text.  We
don't want a situation where we store a statement in a "statement"
variable and write things like "statement.statement"

This swaps out statement for text, so lookups will be statement.text