Define mailing list addresses statically
Previouslly, I was trying to be clever and compose mailing list
addresses programmatically, so that we’d only have to configure each
mailing list’s prefix and then use the constant domain to generate them.

This was too clever by half, though; there’s technically no reason
mailing lists have to all be at the same domain. Worse, I was
piggybacking on the MAILGUN_API_DOMAIN config, which is technically
part of the Mailgun API interaction and has nothing ot do with mailing
lists. (It’s just incidentally the same.)

Unwinding all of this, risking a little un-DRYness (WET?) for the sake
of accuracy and explicitness.

Affects #23