Remove speaker model
Speakers were not actually being normalized in any way, so having
a separate speaker model / table was a bit disingenuous.  This removes
that and simply has speaker name and affiliation as attributes of
a claim.  Some day we can go back.

This migration is fairly complex because it migrates data in addition to
modifying the table structure.

Issue #86