Prevent accidentally mailing external lists
We want to be able to create Real Newsletters that are configured with
Real (External) Mailing Lists, but we don’t want to actually blow them
up with emails while testing.

This protects that from happening by only sending to the newsletter’s
configured mailing list if we’re in production or aren’t configured to
go to an internal list (developers or testers). The latter condition is
important so we can locally configure newsletters to go to testers if
we want.

There are also tests.

Affects #92