Improve cleaning of text templates
I simplified the cleaning of text templates _too_ much, leaving behind
the encoded ampersands. (We have custom text templates for the body of
the newsletter, but re-use the boilerplate from the HTML version to
minimize the likelihood of code divergence.) So, brought back the
sequence runner, but in simplified form.

Went ahead and abstracted out the sequence runner to a utility function
too.

Affects #92