Let us specify which newsletter to test send
Previously, `yarn newsletter:send-test` just sent
`HelloWorldNewsletter`. You had to manually edit the script to send
another.

Now we accept a parameter to specify the one you want (but still
default to the Hello World). This has to be whitelisted through a
switch statement because my metaprogramming is falling short of being
able to instantiate new classes from an argument string, but that’s
probably okay.

Affects #92