Rename PrimaryNewsletter to NationalNewsletter
Right now we have two newsletters on our docket: a local one for
a specific geographic region, and a national one that looks at national
sources (e.g. CNN or national interest twitter).

National is US centric naming, since US is not explicitly spoken, but
since we're US centric for now we can stick with this.  It's likely
that we will refactor the naming / newsletter templating at this level
before adding global alerts.

Issue #92