Add assertions to newsletter validation
We were previously validating newsletters with a getter that would
return a boolean.  This was fine for checking if a newsletter could be
sent, but it meant that there was no easy way to communicate *why* a
newsletter wasn't ready to send.

This uses assertions to validate that the newsletter is ready to send,
which allows us to better keep track of what exacrtly "failed" in those
tests.

Issue #92