Inline CSS during HTML template rendering
We now pass the rendered HTML template through Juice, which inlines the
CSS styles directly on target elements, making the resulting markup
compatible with a much wider array of email clients.

This makes me excited. Not the bloat it adds to our newsletter markup;
that makes me sad. But the fact that in so little code, it solves so
many email client problems (that we’ve found thus far).

Affects #90