Rename style-inlining function
Per feedback by @slifty, while `inline`/`inlining` can be used as a
verb (at least commonly if not literally), it’s not immediately legible
as such, so avoiding it.

Also while typing about this I’ve spelled it `inling` about five times
so that’s another knock against it.

Affects #90