Return a default value when scraper-running fails
This modifies `AbstractScraper.run()` to return a default/empty array
when the scraping itself fails, and expands upon the debug message to
note that it’s happening.

This allows downstream code to be able to always expect an array
(though we will still want to wrap it in a try/catch block for safety).

Affects #115