Add scrape log migration and model
The scrape log is going to replace the use of Redis for storing scrape
history and outcome.  We used Redis originally because it seemed like it
was going to be a snappier solution, but postgres is plenty fast and the
reality is scrape logs are not ephemeral.

If we actually need to optimize for speed we can do so in a later
iteration.

Issue #101