Add two scraper getter methods
We want to know the scraperName of the active scraper even in the
abstract class (for logging and other kinds of messaging), so we're
adding an abstract getter for that attribute.

Also while adding getters, added a getScrapeUrl method to avoid
accessing what would probably normally be a private variable in any
other typed language.

Finally, this add scraper names to the crawlers (which I forgot are
actually extending the scraper)