Add scrape date horizon
This creates a new environment variable called SCRAPE_DAY_HORIZON which
is the number of days in the past that all statement scrapers should look.

This should not change the behavior of the scrapers themselves, but
rather should be leveraged by the worker pipeline when deciding which
scrapes to request.

Right now there is only one scraper pipeline implemented (cnn
transcripts) and this is implemented in that pipeline.

Issue #136