Add timezone support to scheduler
Our scheduler uses cron-style schedule strings. By default, these
follow the system clock on whatever machine is executing the queue.
This is a problem when you want something to happen at some universal
time no matter what machine it’s run on. The obvious answer is “just
express it in UTC”, but _that’s_ a problem if we’re very US-centric and
want the schedule to remain fixed at a clock time regardless of
daylight savings changes, such that “8am Eastern” means 8am in summer
and winter.

Turns out our scheduler, Bull, supports specifying which timezone your
configuration is in, and then translates that to the host system clock
appropriately.

We’re setting our schedule to be relative to New York City, since it’s
the greatest city in the world, greatest city in the world (work, work).

While I’m at it, I went ahead and set the `EVERY_MORNING` schedule to
the time we want (8am).

Affects #153
Affects #154