Fix reference to null schedule
The actual `Schedules` property for “no schedule” is `NONE`, but we
were referencing it as `None`, causing an `undefined` to be passed to
the cron-parser that Bull uses. Besides being wrong, this was throwing
an error (that we aren’t catching, whoops) that tripped up scheduling.

Even after making this change, I’m still seeing two other `undefined`s
passed to cron-parser, so there’s still more work to be done here.

Affects #150