Split getQueue into two methods
We had misunderstood Bull's concurrency mechanism, and whenever we
generated a new queue we would create a queue with the processor
attached.  This actually meant that every single time we ran getQueue()
we were also creating a new job processor for that queue.

This meant that instead of having one job processor per queue we were
having... many more.

This commit creates two types of Queue factory method:

1. getQueue -- this creates a queue object solely for the
   purpose of inserting new jobs.  Generally speaking this is the method
that will be called.

2. startQueueProcessor -- this starts the processor for a queue registered
   to process jobs.  This should never be called from inside
a queue processor script and should really only be called once.

This commit also changes the template for QueueScheduler objects -- it
didn't make sense for them to each implement an external getQueueFactory
method and then invoke it identically to override an internal getQueue.
We now have getQueueFactory as the abstract method and getQueue sitting
as an implemented method in the Abstract class.

Issue #162