Handle rejected promises in job-scheduling tasks
The job-scheduling tasks run a series of promises. Before this change,
we weren’t catching rejected promises, so a problem with any of them
(1) wasn’t handled gracefully and (2) never exited the process, since
`process.exit()` was called during the first `then()`, which assumes
resolution.

Fixed (1) by adding a `catch()` with error-logging and (2) by moving
the `process.exit()` to a `finally()` that executes in all cases.

(Note: `Promise#finally()` is still a draft and not supported by IE
in-browser, but is supported in Node v10+.)

Affects #150