Invert media queries to default mobile
This is discussed more in the attached issue, but essentially not all
email clients pay attention to media queries. This was a problem
because our mobile styles are more important than our desktop styles,
and in the event that one of those scenarios will be ignored, we want
the desktop styles ignored.

So this inverts our behavior so that the mobile styles are the
defaults, and we apply desktop styles with media queries, rather than
the reverse.

Affects #165