Extract and store transcript source/show on claim
CNN transcript URLs embed the source show as a shortcode on the
transcript URL, e.g. the `cnr` in
`http://transcripts.cnn.com/TRANSCRIPTS/1908/13/cnr.01.html`. Extract
this and store it on the claim.

A future commit will add the map of shortcode to show name and display
the name in the newsletters.

Affects #149