Add CNN show names to newsletters
Now that we’re storing the show shortcode on each claim (86f9eef),
let’s display the show name in the newsletter!

This adds the constant that maps shortcodes to longnames and the
Handlebars helper that performs the mapping. This returns empty if the
claim didn’t come from the CNN scraper (or if we don't recognize the
show name); that’s because not all scrapers will want their source
shown in the newsletter. (E.g., the Twitter scraper will store the
username as the source, but we won’t actually want that displayed.)

I’m not wildly keen that we’re embedding markup in this Handlebars
helper, but it’s kosher.

The text version still generates the version of the source name with
HTML markup; that’s not ideal, but it gets stripped out by our
sanitizer anyway.

Affects #149