Refactor source extraction and abstraction
Per some excellent feedback by @slifty, this substantially changes our
approach to extracting, storing, and displaying CNN show names.

Most importantly, we now scrape the actual show name as text from the
transcript HTML rather than using the show shortcode embedded in the
URL and mapping that to a static list of show names that we maintain.
This has the advantage of ensuring we don’t miss a new or renamed show,
though introduces the risk of typos or markup structure changes.
(Still, a net win.)

This also subtantially refactors the `AbstractStatementScraper`
abstraction. All statement scrapers will want to attach some common
properties (`scraperName`, `source`, and `canonicalUrl`) to the
statements before saving them as claims in the database, but _how_ they
do that will vary between platforms. For example, CNN uses the
`scrapeUrl` as the `canonicalUrl`, while Twitter (which will use the
Twitter API rather than web scraping) will extract the `canonicalUrl`
from the scrape response. This is my attempt to abstractly model that:
the statement object modifier has been moved up the chain to
`AbstractStatementScraper` itself and abstract methods created for
getting these properties.

So, in bullet points:

* Extract CNN show name from transcript page HTML
* Remove extraction of CNN show shortcode from transcript URL (and
  associated tests)
* Remove map of shortcodes to longnames
* Remove Handlebars helpers for converting shortcodes to longnames
* Move the adding of additional statement schema properties to the
  statement objects from `CnnTranscriptStatementScraper` up to
  `AbstractStatementScraper` (which incidentally addresses #128)
* Add `getSource()`/`getCanonicalUrl()` abstract methods to
  `AbstractStatementScraper`

One probable area of concern is the use of
`CnnTranscriptStatementScraper.setSource()` and `this.source`, which is
not modeled on `AbstractStatementScraper`. I'm hitting an OOP design
wall and could use some input. I'm leaning toward modeling a Statement
object generally, applying that schema (as properties only, with null
values) to `this.schema` in the `AbstractStatementScraper` constructor,
and then letting `set{Property}()` setters modify that object. Feedback
welcome!

Affects #149 Affects #128

Co-authored-by: Daniel Schultz <slifty@gmail.com>