Collapse repeated whitespace to a single space
Because of line breaks and formatting, we often end up with claims full
of multiple spaces between words. This gets visually collapsed to a
single space in most rendering situations (like HTML), but we shouldn’t
rely on that, and we should clean these down to single spaces when
possible.

This adds a generic `squish()` utility (named after Ruby’s), plus a
scraper utility for iterating over statements and squishing their text,
plus tests, and then adds that squish utility to the
`CnnTranscriptStatementScraper` sequence of statement-cleaning
functions.

Affects #106