Try to enforce immutability of scraper’s scrapeUrl
Per feedback by @slifty:

The scrape URL is the single input parameter when creating a scraper,
and is never intended to be redefined. If JavaScript had immutable
properties, this would be one.

Since it doesn’t, this adds documentation about our immutable intent,
and then further attempts to enforce that with a `setScrapeUrl()`
helper method that throws an error if the scrapeUrl is changed while
already defined.

Of course, a dev could simply ignore the documentation and the helper
and change `this.scrapeUrl` directly, but who let that person in?

Affects #149