Allow scrapers to specify authorization headers
When scraping cnn we didn't need to worry about authorization or HTTP
headers.  This will not be the case with Twitter, so we need the
AbstractScraper to allow implementing classes to inject headers into the
request.

There was some consideration of modifying the abstract scraper to
function in two modes, using request or oauth depending on whether the
scraper relies on oath.  This punts that decision for later by allowing
the twitter scraper to insert oauth headers directly.

Issue #26