Add the TwitterStatementScraper
The TwitterStatementScraper takes a twitter handle and converts that
user's most recent tweets into Statements.

For now we are not normalizing the user names according to the
internally managed spreadsheet (e.g. if a user changes their twitter
name it will result in a different statement / claim)

This commit also does not invoke the twitter scraper as part of the
scraper pipeline.

Issue #26