Add twitter accounts table
Our twitter scraper is not going to be scraping everything on twitter,
so we need a mechanism to specify which accounts are of interest.

This commit creates the migration and model for tracking which accounts
are which.  The list_name column allows us to have more than one
"bucket" that the accounts are in (which will be needed for the short
term as we have localized newsletters)

This table will be populated in a separate scrape of the existing google
spreadsheets for the two twitter lists; eventually we may want to
explore a local administrative interface for that list management.

Issue #26