Add statement log model
This creates the concept of a statement log, a schema responsible for
storing when statements have been seen by the system.  This will allow
us to have our pipeline avoid re-processing statements that had been
scraped before but did not result in a claim.

It is important to note that this log will only produce more granular
information than the scraper log for scrapers in which statements have
specific IDs within a scrape (e.g. twitter, where each tweet has its own
ID.)  For scrapers in which the scrapeUrl is the same as the
canonicalUrl for all resulting scrapes this log will provide the same
information as the scraperLogs.

This commit includes a new migration, so you should run

`yarn migrate`

after merging.

Issue #26