Log statements after scraping
When a new statement is scraped, this will now insert a log entry for that
statement.  Note that a log entry will not be created if the statement
had been previously logged.

It also looks up the most recent log entry (before inserting
a log, obviously, since otherwise we would just have the log entry we
just inserted) and attaches that information to the statement so future
pipelines can know if the statement had been seen before and when.

Furthermore, this changes the footprint of scrapeHandler by making it an
async method. This was necessary because statement scrapes need to look
up scrape logs to know if the scrape had been seen.  I went and made the
AbstractCrawler async too although I'm realizing now that might not be
necessary.

Issue #26