Add env-settable support for ClaimBuster API V2
This adds optional support for V2 of the ClaimBuster API, which is
enabled by setting the appropriate `.env` variable.

The new API has a different URL structure and requires an API key, so
this also adds two `ClaimBusterClaimDetector` class methods to help
compose the response and smooth out the differences. (I struggled with
whether these should be utility functions, but decided in the event we
wanted to do things like abstract out ClaimBuster integration into a
base class for calling other endpoints, it belonged here.)

This also moves the `CLAIMBUSTER_THRESHHOLD` from the shared constants
file into a `claimDetector`-specific one. This may be a move in the
wrong direction, but at least it's consistent with our other
compartmentalization of constants.

Resolves #139