Convert ClaimBuster API integration to V2
This changes our ClaimBuster integration to interact with V2 of the API.

We now POST the text to the API in a JSON payload rather than just
appending the statement to the URL as a GET request. Not only is this
more correct, it has the side effect of resolving the issue with
slashes in content (#204), which was due to ClaimBuster’s GET requests
failing on encoded slashes (`%2F`), without requiring us to strip them
out (#214). Thus, this also removes the now-unnecessary text-cleaning
and associated tests.

Since the V2 endpoint requires authorization, this adds a new `.env`
variable for the API key that must be set before integration will work.

Resolves #139
Resolves #204
Resolves #214