Compose Twitter API URL query string w/querystring
Composing the Twitter API URL query string parameters using direct
string manipulation is error-prone, creates a really long text line,
and isn’t quite as legible as composing it as an object. Since Node
comes with `querystring`, let’s use that to compose it.

This resolves no actual issue, but is being committed as part of #213.