Add a generic scraper
This creates a generic scraper object which will simply return the raw
content that is pulled from an HTTP request.  This is going to be needed
for an incoming utility script that will make it possible to generate
automated tests for scrapes.

Without this there is no way to scrape, for instance, a CNN transcript
without also invoking the entire statement extraction pipeline.