Add cnn test generation utility
The CNN scraper is a full blown content manipulation pipeline, and
issues with the final result, while relatively to address and debug, are
hard to translate into a meaningful test that respects the full
pipeline.

Ultimately we want to be able to write a test that follows each step of
the transformation from a challanging transcript.  This utility script
will take in a transcript URL and then output each transformation step
into a format that automated tests can later pick up.

To run the script you type:

yarn generate:test:cnn -u http://transcipt.cnn.com/continuehere

Issue 127