Invoke meta tests for cnn pipeline
This modifies the cnn automated tests to actually invoke the pipeline
steps using the inputs and outputs from our test generation script.

Simply put, this will iterate through the instances of input / output
pairs and invoke the relevant utility method.  If the output of the
utility method does not match the expected output, the test will fail.

Issue #127