Remove honorifics using improved pattern-matching
Previously, we removed honorifics from the speaker name using a form of
`replace()` that didn’t ensure the honorific appeared on its own
(either rooted at the beginning of the string and followed by a space,
or surrounded on both sides by a space). This made it unusable for
honorifics like “MAYOR” that could likely appear within someone’s name
(like “SONIA SOTOMAYOR”).

This both improves upon that and lists the honorifics in an
easily-appendable array, as well as adds trickier tests around them.

Resolves #186