Revise chunk break RegEx
The CNN parsing pipeline uses a series of regular expressions to do
things like break a long transcript string into isolated chunks.  We
were seeing some attribution errors related to a mistake in the way
those chunks were identified.

This adjusts the RegEx to be hopefully more universal, handling the
reality that CNN sometimes includes lowercase words and letters in the
attributuon section of a statement chunk.

This also begins to move the regex into smaller re-usable components.

Issue #167
Issue #64