Allow periods in CNN affiliations
Attribution affiliations sometimes include periods (e.g. if the title
contains the word "U.S." so we need to allow that.

This also exposed a secondary bug related to the way statements were
being extracted from chunks.

The CNN statement pipeline had logic to remove attribution clauses from
chunks; the issue is that in a very rare case it would also additionally
remove segments of the statement itself.

Since attributions will only ever be at the beginning of a chunk, we can
safely add the start-of-line anchor to the attribution regex when
deciding what to remove.

Issue #167