Clean up mid-statement descriptor cruft
Sometimes a CNN transcript might have a voice-over indication in the
middle of the statement (e.g. to represent a visual transition without
a speaker change).

In some circumstances this would result in a rogue colon surrounded by
whitespace.

Before recent changes, that rogue colon caused attribution errors.  That
is no longer the case, but the colon is still cruft that doesn't belong.

This commit removes that colon (leaving the whitespace to be squished
later).

Issue #167