Move parentheticals to affiliation
CNN transcripts would sometimes have parenthetical notes as part of
an attribution, usually to indicate things like party affiliation.

Depending on the location, our logic would often incorporate this
block with the name, rather than affiliation.

Those parentheticals are not actually an aspect of a name (arguably they
also are not an aspect of affiliation, but that is not a bridge to cross
right now).

This commit parses those components to the affiliation field and cleans
up the name to remove any resulting whitespace (since the preivous
delimiter was always a , there was no chance of whitespace before).  It
also modifies the affiliation trim so that the opening parentheses would
not be removed if it exists.

Issue #167