Remove parenthetical consideration from cleaning
Names used to sometimes contain parentheticals when parsing, and as
a stopgap we avoided including them by stripping them out.  Now that
we're properly putting those in the affiliation portion of an
attribution we no longer need to clean them from name (since they won't
be in names going forward).

This removes that line and the associated test.