Increase length of scrape_logs.scrape_url field
Previously, we defined the `ScrapeLog.scrapeUrl` attribute (in Postgres
terms, `scrape_logs.scrape_url`) using the Sequelize `STRING` shorthand
with no length, which meant it defaulted to 255. As with
`StatementLog.canonicalUrl`, a more appropriate length for URLs is 1024.

This adjusts the model definition and adds a Sequelize migration to
change the database column. Consequently, you should run migrations
after merging this commit.

Affects #242