Match up Claims table and model definitions
This commit adds a migration to change the length of three `claims`
table columns and modifies one `Claim` model attribute definitions to
match the existing column (which was, incidentally, more correct).

Previously, we had a mismatch between our Sequelize model definition
for claims and the actual table structure. (That this is possible
belies a bit of a problem in Sequelize itself, but I digress.) Four of
the fields were defined as 1024 characters long in the model, but used
the default 255-character length in their original migrations.

One of those fields, `scraper_name`, didn’t actually need to be 1024
anyway, so this changes the model definition. The other three could
arguably fit into less than 1024 in many cases (especially `source`),
but there’s no harm in making them long enough for edge cases.

You will need to run migrations after merging this change.

Affects #253

Signed-off-by: Justin Reese <justin@justinreese.com>