Add support for `Claim.claimedAt`
We’ve had a `Claim.claimedAt` attribute, but haven’t been using it. In
preparation for allowing Twitter (#258) and CNN (#100) to store the
actual time a claim was published rather than just when we scraped it,
we need to have these components ready:

- Have `AbstractStatementScraper` fall back to the current timestamp if
  the scraper didn’t set one
- Have the claim detector pass `claimedAt` all the way through to the
  claim-saving
- Use `claimedAt` rather than `createdAt` for newsletter composition
  queries

This is the infrastructure to allow populating the field; another
commit will actually populate it.

Affects #258
Affects #100

Signed-off-by: Justin Reese <justin@justinreese.com>