Add active/inactive concept to TwitterAccount
We periodically update the Twitter accounts we want to scrape. Rather
than deleting them from the database, we want to be able to deactivate
and reactivate them, and we want the scrape and newsletter composition
queries to scope by active state.

This commit adds the fundamental concept of `TwitterAccount.isActive`,
including its database migration (and column configuration) and scoping
relevant queries.

Affects #255

Signed-off-by: Justin Reese <justin@justinreese.com>