Add script for deploying the service
This script will deploy the service on a properly configured server.
The server needs to have pm2 installed and configured, though some day
we might want to adjust that.

The intent is that this script would be invoked by our CI service after
the deliver.sh command is run.

Issue #275