Set up continuous deployment for dev
This updates our travis configuration so that commits to the master
branch of the repository are automatically propogated to our dev server.

This requires a few new travis configuration values; in particular:

- STAGING_SERVER :: the hostname of the staging server.
- STAGING_SERVER_USER :: the username to use when deploying to staging.
- STAGING_BRANCH :: the branch that should trigger continuous deployment.

This also created an encrypted deployment key that can be used to deploy
to the dev server.  The public portion of the key has been added to
deploy_key.pub and that can be added to any server we want to be able to
deploy to via travis.

The deploy key was set up using the steps outlined in [this
guide](https://github.com/dwyl/learn-travis/blob/master/encrypted-ssh-keys-deployment.md)

Issue #275