Remove identity flags
We may want to add identity back some day so this script is easier to
run locally, but for now that isn't a primary use case and it was
causing issues as-implemented within the primary use case (e.g. when
being invoked on Travis).

This removes the code supporting the explicit passage of an identity
file in the travis script.

Issue #275