Update parcel to 1.12.4
This patch primarily is being made to get rid of the security
vulnerability notification on github.  The vulnerability didn't actually
affect our code base since the relevant method in the sub-library was
not being invoked by parcel.  Nevertheless, security warnings look scary
and are worth addressing even if they are inaccurate!

Issue #285