Add known speaker scraper queue
The known speaker scraper queue invokes the google scraper to extract
a list of known speakers and store that list in our database for use by
the system.  This involves a new environment variable:

- KNOWN_SPEAKERS_GOOGLE_DOC_ID

This variable contains the google doc ID of the known speakers doc.  The
doc should have a first_name and last_name column.

Issue #281