Fix Twitter scraper to normalize speaker name
In #268 and #294, we updated the Claim model to split `speakerName`
into two fields, (`speakerExtractedName` and `speakerNormalizedName`).
Our speaker name normalizers are supposed to store the original
extracted name in both fields, then normalize the name and store it in
`speakerNormalizedName`.

We made this change for CNN, but not for Twitter. This commit covers
Twitter.

Resolves #298

Signed-off-by: Justin Reese <justin@justinreese.com>