Add known speaker filter to national newsletter
This modifies the way claims  are looked up for the national
newsletter so that claims made by known speakers are prioritized.  The
logic is to attempt to fill the entire claim list with recent
known-speaker-claims.  If there are not enough claims by known speakers,
the query will then return claims made by other speakers.

There was a lot of discussion on the best way to implement this query.
The truly best approach would be to leverage sequelize's "through"
keyword in the query builder.  To do this, however, we would need
an explicit association between claims and known_speakers.  That felt
too heavy to implement today, since that association would imply
a level of fidelity that the system simply does not provide.

For instance, we do no form of identity resolution; rather, we are
simply looking at the names to determine if a speaker is a "known
speaker".

We also considered the creation of a view, instead of an ad hoc query.
Since Sequelize does not officially support views, this felt like it
would manifest as a heavier version of technical debt (it would require
a migration and a new "model" for the view.)

This felt like the lightest solution for today; we should revisit
when we are ready to build out identity resolution / formally store
speakers as separate entities from claims.

Issue #280