Remove HelloWorldCrawler
When we first created the concept of crawlers and queues we created
a Hello World crawler (and queue) to show how our worker / queue
architecture should work.  Now that we have real queues, that is no
longer needed in this project and these are now just cruft sitting in
the code.

Unlike the HelloWorldNewsletter there is no clear value in having the
crawler.

Issue #312