WIP - Add CnnTranscriptStatementScraper tests
This adds a mocked test for the CnnTranscriptStatementScraper component.
This invokes the jest snapshot feature which is similar to the CNN util
test suite functionality we built manually before.  What this does not
include is tooling for adding additional transcript mocks, which is
a bridge we can cross some day.

This is a WIP because the tests hang currently.  This is because the
CnnTranscriptStatementScraper is the first component we have tested
which interacts with the database, and as a result we don't have the
proper tooling to close the database conneciton.  There is an open issue
(#317) which will address that, at which point we can remove this note
and the WIP and merge this test to master.

Issue #309